Secure Cloud Computing/Ropetyapp-devsec-main/requirements.txt
Flask==3.0.0
bcrypt==4.1.2
PyJWT==2.8.0
google-cloud-storage==2.14.0
pymysql==1.1.0
google-cloud-secret-manager==2.18.0
SQLAlchemy==2.0.23
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
requests==2.31.0
pyotp==2.9.0
qrcode[pil]==7.4.2
google-cloud-logging==3.8.0
